+++ releasenotes.md
@@ -67,6 +67,7 @@
 - @tag didn't work with members #2236.
 - Assert comparing untyped lists #2240.
 - Fix bugs relating to optional interface addr-of #2244.
+- Compiler null pointer when building a static-lib with -o somedir/... #2246
 
 ### Stdlib changes
 - Deprecate `String.is_zstr` and `String.quick_zstr` #2188.